110 ms (5.06%) 87.8 MB (5.28%) 0224-basic-calculator .java Hard Math

+++ LeetCode/Hard/0224-basic-calculator/0224-basic-calculator.java
@@ -45,7 +45,7 @@ class Solution {
 
     }
     public String cal(Stack<String> stack){
-        int num=0;
+        long num=0;
         String op="*";
 
         while(!stack.isEmpty()){
